% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmygrad-package.R
\docType{package}
\name{stanmygrad-package}
\alias{stanmygrad-package}
\alias{stanmygrad}
\title{The \pkg{stanmygrad} package.}
\description{
Examples of custom gradient implementation in Stan.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
